pyproject.toml
[tool.poetry]
name = "justhost-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "justhost_app"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.5"
django-rest-framework = "^0.1.0"
flake8 = "^7.1.1"
django-filter = "^24.3"
psycopg2 = "^2.9.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
